requirements.txt
flask
flask-socketio
python-dotenv
gunicorn
eventlet